Show current playing task after favorites but before other ones
e.g. you are playing music with Spotify, then you open other apps
so Spotify becomes the 15th app in the recents list.
You'd have to scroll the panel up to find spotify and open it if needed.

With this commit, instead, if Spotify is playing, the panel will always
show it in the first positions, after top task, last expanded apps and
favorite apps.
When paused or stopped, the task will get its original position in the
list.

When the player is playing there's a high probability that you'll
need to open it so it's good to have it soon without the need to scroll
up the panel.

Change-Id: If1ee9d0299e1588c6ebf9363a38d15881030a584